refactor: move voc_match_cache into _LegacyResourcesMixin; drop dead imports
voc_match_cache is a back-compat accessor for the per-skill _voc_cache
dict; its setter already self-identified as deprecated. Move both
property and setter into the mixin so they disappear with the rest of
the legacy surface, and replace the stringly-typed log.warning with a
real DeprecationWarning.

Also drop now-unused imports from ovos.py: classproperty (only consumer
was runtime_requirements, which moved with the previous commit) and
isdir (only consumer was the hand-rolled locale-dir walk, now folded
into iter_locale_dirs).
